feat: useAuthQuery 를 수정한다
* mutate 콜백의 위치 변경
* 우리가 정한 코드 컨벤션에 맞도록 파일명, 함수명, 변수명을 수정했다.
* 타입의 범위를 명시적으로 줬다. 좀 더 좁힐 수 있을 것 같지만 getApiInstance()  와 의존성이 있어보이는 작업 같아 제외했다.
